Separate out playbook variables
Separated out all variables from the playbooks.
Not sure if this is the best approach, simply including
the variable files in the playbook where needed.

There is a group_vars and host_vars we could also use to
configure variables.

Also, split variables between regular and sensitive ones.
The secrets.yml file is not put into the repository.
Later we may add in script to generate all secrets at
random before deployment.