Gather cloud manager key and inject into agents
Ths commit implements downloading the key from
the cloud management server and adding it to
all cloud agents.  This allows the cloud manager
ssh access to all agents, which is needed
by management UI for initial tasks setting up
the agent.  We also add in the manager
ip to the hosts.allow to allow for ssh
access from the manager.