First attempt at making redeploy smoother
We may later fix individual tasks that should not be
done on a subsequent deployment.  But for now define
a variable that is true only on initial deploy.
If you set `deploy` to false then those tasks
that should not be done on a second subsequent
redeploy will be skipped.

You can specify this on the command line as an
extra variable like:

$ ansible-playbook deploy-cloudstack.yml --extra-vars '{deploy: false}'