Reimplement initial ansible ssh key injection in ansible
Reimplement initial ansible ssh key injection to use
ansible to perform the task.  We use the
authorized_key module and temporarily set
the ansible user and password to be the
throway user.  We use the ability to
become root and that the cloudstack user
is created with sudo privileges to complete
the tasks.