Fix deployment bugs and start on post deployment setup
Fixed a bug in deployment where playbook stops when
trying to delete the temporary insertion cloudstack
account.  We are not killing any hanging on cloudstack
processes.

Started on post deployment setup work.  Mainly we
are now installing cloudmonkey command line tool in the
post setup, and we have an example of a script file
that attempts to set up all of the basic network
infrastructure now using the cmk tool.