Configuration of https/ssl certificates and access for cloudstack
This commit represents a significant enhancement to configure
https/ssl access for the cloudstack management api.  Most
significantly, many tasks have been added into the setup-cloudstack.yml
playbook that create a self signed certificate and configure the
management ui to enable https.  The certificates are self-signed,
but work to allow access to the management api by tools
like cloud monkey and the ansible cloudstack library tasks.
The certificates allow access to the cloudstack management UI,
though we still get access warnings that have to be ignored
in firefox with this self signed certificate.