Example provisioning of infrastructure using ansible cloudstack module
This commit has some significant progress in an example of how to
use the ansible cloudstack module to provision cloud infrastructure.
The main example is found in the setup-infrastructure.yml
playbook.

In general ansible cloud configuration looks like it will be very
useful.  This may work as well, if not better, than hot scripts
from OpenStack to automate creating types of infrastructure on
a cloud.