Create admin api key and secret when setting up cloud
As we had guessed, the admin api key and secret do need
to be explicitly generated before we can use them with
cmk and ansible cs tools.  We generate the
key and secret now before inserting into the
configuration file.

We also rearrange the order of the script a bit
to try and avoid some timing issues.  After we
restart the cloudstack-management service it
takes some time before the api ports become
available for queries, so move down some code
that attempts to access this later on.  This
maynot be sufficient, perhaps we need to
explicitly wait until management api ports
are up before continuing?