Add in additional revers proxy configuration examples
Expand nginx reverse proxy server configurations for
some additional servers on the cloud network.

This playbook and configurations still have some things that need
to be done.  Primarily, the two new configuration in this
playbook use copies of the certificate and key we
have for the cloudstack server.  These are not copied
by the playbook.  Should either generate new ones here on
this server for these vhosts, or else copy them in
the playbook.