Update and fix ssl/https certificate generation procedures
We have revamped our ssl certificate generation.  The updates
use the following site discussion, which seems to be
working much more smoothly for our cert creation:

https://devopscube.com/create-self-signed-certificates-openssl/

Basically following that example, we create our own
root CA and then self sign server certificates.  The
configuration of the csr and cert properties is what is
most likely the key we needed to smoothly generate
our own root CA for self signing.

Our procedure now has a root CA certificate generated on
the ansible server that we use for all self signing.
Then we create csr for each needed server and sign them
with our ansible root CA.  This procedure is successful
in allowing firefox and other browser access if we
add the root CA we used into the browser list of
authorities.