Attempt at using ssl for console proxies
Tried to add support for ssl to console proxies.  The following
seems to work, and url when open console proxy says it is
TLS encrypted.  But I was expecting to use a domain
name in url like 192-168-000-154.consoleproxy.harter.priv,
but I still see the ip address.

We did the following:
We updated the hosts file to include the
xxx-xxx-xxx-xxx.consoleproxy.harter.priv mappings and put on all hosts
in
the cloud.

Then uploaded the key.  Used both cmk, and also used the management UI
to try and upload:

  195  openssl pkcs8 -topk8 -in cloudstack.key -out cloudstack.pkcs8.encrypted.key
  196  openssl pkcs8 -in cloudstack.pkcs8.encrypted.key -out cloudstack.pkcs8.key
 (localcloud) 🐱 > upload customcertificate domainsuffix=cloud01.harter.priv id=1 name=root certificate=@HarterRootCA.crt
 (localcloud) 🐱 > upload customcertificate domainsuffix=cloud01.harter.priv id=2 name=server privatekey=@cloudstack.pkcs8.key certificate=@cloudstack.crt