Add in generation of kickstart files from profiles
Added in capability to generate the a kickstart file
for a host based on the target profile selected when
registering the host.

Basically a profile is a user-data file that we will
generalize to autoinstall a class of machines.  For example,
probably all compute nodes can use the same profile if we
generalize it so that it uses all of the hard disk
when partitioning the disk.  Probably need separate
profiles for the manager and storage nodes, e.g. will
probably need to specifically setup storage devices
as needed on those machines.

Also tried to add in more capabilities to the user-data
files.  We now can inject the ansible management
ssh key using this file for the cloud-init.  We
also do some other things like set up sudo
no password, and configure even the initial network
to use static ip.