Add in yaml configuration file
Doing some refactoring.  Not sure if want to stick with using
yaml file for global settings / configuration.  But in this
commit we have pulld out the configuration, put it into
a yaml file, and added a mechanism to have the global
settings read in from the file so they can be accessed
in the script.