Create playbooks for some example/useful servers
Began working on creating a collection of playbooks to
setup useful examples of servers.  So far this
commit has a revers-proxy, plex and basic lamp
server.

These types of playbooks offer similar/same functions
as hot scripts we used to use on OpenStack.

These 3 server configs all have a common start
to create a basic Ubuntu 22.04 server.  We are planning
on refactoring

- Use roles
- Create a role for making a basic ubuntu 22 vm
- Then have roles for servers types we want to
  support.