Update default account name used for management
We are going to use "cloudmanager" as the expected
account name that needs to be sshable into by ansible
using the management key.  We are making this change
mostly so that we can also setup the ansible
manager host itself to have an account of the
same name so we can also perform ansible tasks
through that account.